pyproject.toml
[project]
name = "dq"
version = "0.1.0"
description = "Training data quality detection agent for LLM pre-training and SFT fine-tuning"
readme = "README.md"
authors = [
    { name = "Peter Zheng", email = "peter@quantforge.dev" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "datasketch>=1.6",
    "xxhash>=3.0",
    "regex>=2023.0",
    "tqdm>=4.60",
    "rich>=13.0",
    "openai>=2.29.0",
    "anthropic>=0.86.0",
    "datatrove[io]>=0.4.0",
    "spacy>=3.8.14",
    "zstandard>=0.18",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
lid = [
    "fasttext-numpy2-wheel>=0.9.2",
    "fasteners>=0.20",
]
model = [
    "fasttext-wheel>=0.9",
    "transformers>=4.30",
    "sentence-transformers>=2.2",
    "torch>=2.0",
    "openai>=1.0",
]
bench = [
    "datasets>=2.14",
    "zstandard>=0.18",
]

[project.scripts]
dq = "dq.cli:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=7.0", "pyarrow>=14.0"]
